django_API/requirements.txt
Django==1.7.4
djangorestframework==3.0.4
wsgiref==0.1.2
